--- src/components/Badge/index.ts
@@ -1,1 +1,0 @@
-export { default as Badge } from './Badge';
