+++ src/stories/Badge.stories.tsx
@@ -1,0 +1,18 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { Badge } from './Badge';
+const meta: Meta<typeof Badge> = {
+  title: 'UI/Badge',
+  component: Badge,
+  args: { count: 1, max: 999, showZero: false },
+  argTypes: {
+    count: { control: 'number' },
+    max: { control: 'number' },
+    showZero: { control: 'boolean' },
+  },
+  parameters: { layout: 'centered' },
+};
+export default meta;
+
+type S = StoryObj<typeof Badge>;
+
+export const ShowZero: S = { args: { count: 0, showZero: true } };
